feat: restore browser panel refactor
